requirements.txt
asgiref==3.8.1
awscli==1.36.17
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
docutils==0.16
idna==3.10
jmespath==1.0.1
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.3
